* next: (26 commits)
  minor, process BIDASK ticks in ute-shnot
  minor, track the smallest snapshot interval for SSNP_FLAVOUR ticks
  minor, only use 8U wide intv blobs as we only support info-tracking of bid ask and trade candles
  minor, use intv slots 4U and 5U to keep track of snapshot intervals
  fix, PEBKAC, a SL1T_TTF_BIDASK candle *is* of SSNP_FLAVOUR, remove dupe
  minor, provide an interval tracking system based on a scale table
  fix, PEBKAC, for non-mremap() case in resz_gbs() actually munmap() the old array
  minor, make ute info traverse the whole file instead of just the first page, and use gbs for bitsetting
  minor, import truffle's gbs system for generic bitsets
  minor, actually print a timestamp when using ute-info -i|--interval
  minor, introduce -i|--interval and -m|--modulus in ute-info for per-period information
  minor, refactor bitset printing in ute-info
  chore, PEBKAC, list info.5.ref.* in list of files to distribute
  test, provide more regressions for ute info tool
  minor, flip opposite-endian files passed to ute info automatically
  chore, for BIDASK ticks in info mode print tick_q (as in a full quote)
  test, provide regressions for ute info tool
  hygiene, replace printf()s without actual formatting by their fputs() equivalents
  minor, also provide info output for ticks of BIDASK type
  hygiene, cast index slot for ute_bang_symidx() in pump_slut()
  ...

+++ src/ute-slut.c
@@ -253,7 +253,7 @@ pump_slut(int outfd, utectx_t hdl)
 		}
 		/* time to bang! */
 		if (*sym) {
-			ute_bang_symidx(hdl, sym, idx);
+			ute_bang_symidx(hdl, sym, (uint16_t)idx);
 		}
 	}
 	munmap(buf, bsz);
